refactor: split user service into queries and commands
- Delete user.service.ts (god file with mixed concerns)
- Create user.queries.ts: getDashboardStats, getProfile, findUser
- Create user.commands.ts: updateProfile, updateContacts, banUser, unbanUser
- Add USER_PROFILE_SELECT const to eliminate field selection duplication
- Add handleNotFound duck-type P2025 helper (avoids Prisma enum import)
- Inline type annotations for groupBy results (role, status, userAgent)
- Update user.router.ts and ai.router.ts to import from specific files
